backend/requirements.txt
# Core Dependencies
PyGithub==2.8.1
convex==0.7.0
python-dotenv==1.2.1
fastapi
uvicorn[standard]
email-validator

# HTTP Clients
httpx==0.28.1
requests==2.32.5

# LangChain and AI
langchain-google-genai
langchain-openai
langchain
langchain-core
langchain-mcp-adapters

# Moss and Embeddings
inferedge-moss==1.0.0b5
inferedge-moss-core==0.1.5
transformers==4.57.1
tokenizers==0.22.1
onnxruntime==1.23.2
safetensors==0.6.2

# Utilities
agentmail
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
coloredlogs==15.0.1
cryptography==46.0.3
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
mpmath==1.3.0
numpy==1.26.4
packaging==25.0
protobuf==6.33.0
pycparser==2.23
PyJWT==2.10.1
PyNaCl==1.6.0
PyYAML==6.0.3
regex==2025.10.23
sniffio==1.3.1
sympy==1.14.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0